@@ -15,25 +15,32 @@
     "require": {
         "caseyamcl/toc": "~1.0",
         "composer/semver": "^1.4.2",
-        "doctrine/annotations": "^v1.2",
         "erusev/parsedown": "^1.6.3",
         "ext-bcmath": "*",
-        "fiasco/metayaml": "^1.5",
+        "ext-ctype": "*",
+        "ext-iconv": "*",
+        "fiasco/metayaml": "^2@dev",
         "fiasco/symfony-console-style-markdown": "^1.2",
         "guzzlehttp/guzzle": "~6.0",
         "kevinrob/guzzle-cache-middleware": "^3.2.1",
         "mustache/mustache": "^2.11",
         "paragonie/random_compat": "^2.0",
-        "php": "^7.2",
-        "symfony/cache": "^4.0",
-        "symfony/console": "^4.0",
-        "symfony/expression-language": "^4.0",
-        "symfony/filesystem": "^4.2",
-        "symfony/finder": "^4.0",
-        "symfony/process": "^4.0",
-        "symfony/validator": "^4.0",
-        "symfony/yaml": "^4.0",
-        "twig/twig": "~1.0"
+        "php": "^7.2.5",
+        "symfony/console": "5.0.*",
+        "symfony/flex": "^1.3.1",
+        "symfony/cache": "^4.4|^5.0",
+        "symfony/config": "^5.0",
+        "symfony/dependency-injection": "^5.0.1",
+        "symfony/error-handler": "^4.4.1|^5.0.1",
+        "symfony/http-foundation": "^4.4|^5.0",
+        "symfony/http-kernel": "^5.0",
+        "symfony/polyfill-mbstring": "~1.0",
+        "symfony/filesystem": "^4.4|^5.0",
+        "symfony/finder": "^4.4|^5.0",
+        "symfony/yaml": "5.0.*",
+        "twig/twig": "3.0.*",
+        "symfony/process": "^5.0",
+        "symfony/expression-language": "^5.0"
     },
     "conflict": {
       "drutiny/http": "^2.0"
@@ -46,8 +53,12 @@
     },
     "bin": ["bin/drutiny", "bin/build_phar"],
     "require-dev": {
-        "phpunit/phpunit": "^5.0",
-        "squizlabs/php_codesniffer": "2.*",
-        "drupal/coder": "^8.2.0"
+        "phpunit/phpunit": "*",
+        "squizlabs/php_codesniffer": "3.*"
+    },
+    "scripts": {
+        "auto-scripts": {
+
+        }
     }
 }
