@@ -4,7 +4,7 @@ namespace Drutiny\AuditResponse;
 
 class NoAuditResponseFoundException extends \Exception {
   protected $name;
-  public function __construct($name, $message)
+  public function __construct(string $name, string $message)
   {
     $this->name = $name;
     parent::__construct($message);
