@@ -2,16 +2,13 @@
 
 namespace Drutiny\DomainList;
 
-use Drutiny\Target\Target;
-
 interface DomainListInterface {
-
-  public function __construct(array $metadata);
 
   /**
    * @return array list of domains.
    */
-  public function getDomains(Target $target, callable $filter);
+  public function getDomains(array $options = []);
+  public function getOptionsDefinitions();
 }
 
  ?>
