@@ -18,8 +18,8 @@ class JSON extends Format {
    */
   protected $output;
 
-  public function __construct($options) {
-    parent::__construct($options);
+  public function __construct() {
+    parent::__construct();
     $this->setFormat('json');
   }
 
